refactor: MockExchangeRate 위치 이동
