+++ src/test/java/shopping/config/TestOtherApiConfiguration.java
@@ -4,7 +4,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
 import shopping.order.service.ExchangeRateService;
-import shopping.other.exchange.MockExchangeRateServiceImpl;
+import shopping.other.exchange.currencylayer.MockExchangeRateServiceImpl;
 
 @Configuration
 @Profile("test")
